+++ tests/testthat/test_methods.R
@@ -8,4 +8,5 @@ test_that("neutrality test identifies neutral and non neutral examples", {
   expect_true(neutralitytest(VAFneutral)$area$metric < neutralitytest(VAFselection)$area$metric)
   expect_true(neutralitytest(VAFneutral)$Dk$metric < neutralitytest(VAFselection)$Dk$metric)
   expect_true(neutralitytest(VAFneutral)$meanDist$metric < neutralitytest(VAFselection)$meanDist$metric)
+  expect_true(neutralitytest(VAFneutral, read_depth = 100.0, cellularity = 0.8)$meanDist$metric < neutralitytest(VAFselection)$meanDist$metric)
 })