+++ NEWS.md
@@ -1,3 +1,7 @@
+# neutralitytestr 0.0.2
+- Added option so that maximum of integration range can be calculated from read read depth
+- Added mu/beta = annotation to lsq_plot rather than y = 
+
 # neutralitytestr 0.0.1
 
 First realease of package.